MainPage fully working
