+++ src/Context.js
@@ -1,15 +1,12 @@
 import React, { useState } from "react";
 import PropTypes from "prop-types";
 import FetchData from "./services/FetchData";
-/*eslint-disable */
-import store from "./redux";
 
 const Context = React.createContext();
 
 function ContextProvider({ children }) {
   const [isLoading, setIsLoading] = useState(false);
   const [isError, setIsError] = useState(false);
-
   const [setUrl] = FetchData();
 
   const getNum = (url) => {
